Version 0.06 - Improved UI & Caching
The UI has been slightly improved to give more space to the actual 'browser' (web page).

A Caching system has been added to improve performance. Here is how it works:

Cache Storage: A dictionary self.cache is used to store the HTML content of visited pages. (This is saved LOCALLY on your device)

Cache Logic: Before loading a URL, Stratus checks if the URL is in the cache. If it is, the page is loaded from the cache.

Cache Expiry: A timer is set to clean up the cache every hour, removing entries older than one hour.

Lazy Loading has also been added to improve the users experience! Here is what it does:

Lazy loading means that the content of a tab is only loaded when you actually click on that tab to view it.

Why it's useful: This helps save memory and resources because the browser doesn't have to load all tabs at once. It only loads what you're currently looking at.

How it works:
The load_tab_content method checks if the tab's URL is valid.
If the URL isn't valid, it sets a default URL (like "https://www.google.com") for that tab.
This method is called whenever you switch to a different tab.