Version 0.07 - Favicons & Updated UI
Updated UI:
Moved the New Tab button to the nav bar line instead of being up with the tabs, this makes for a much cleaner look. The tab text is now bold for a more modern look!

Favicons are here! That's right, I am introducing favicons, which work with ALL websites! (They are a little buggy but I will fix this inn a future update, probably the next one or two)

Changed some Keyboard shortcuts to be better (make more sense)