Version 0.08 - Blocking Trackers & other fixes
All known trackers that Stratus knows about will block them from tracking you! Currently there are 123 domains that Stratus will block! This does not mean that the site will be blocked (eg. adobe.com), but it means that the tracker will be blocked on that site.

Fixed a fatal bug! This bug would crash the browser after one hour due to the cache trying to clear itself to save memory.